refactor: move App composable to own file
